Extract module `_cache`
